add win-specific relativize tests
